refactor: move model filters to seperate component
