+++ src/components/ModelSelect/ModelSelect.vue
@@ -8,14 +8,13 @@ import ModelIcon from '../Icon/ModelIcon.vue';
 import { TbListDetails } from 'vue-icons-plus/tb';
 import isOnMobile from '@/utils/core/isOnMobile';
 import { useModelList, type ModelInfoListItem } from '@/composables/useModelList';
-import { AiOutlineEye, AiOutlineLoading } from 'vue-icons-plus/ai';
+import { AiOutlineLoading } from 'vue-icons-plus/ai';
 import PrimaryButton from '../Buttons/PrimaryButton.vue';
-import { BiBrain, BiFilterAlt, BiRefresh, BiWrench } from 'vue-icons-plus/bi';
-import MultiItemSelect from './MultiItemSelect.vue';
+import { BiFilterAlt, BiRefresh } from 'vue-icons-plus/bi';
 import FloatingMenu from '../FloatingMenu/FloatingMenu.vue';
+import FilterMenu from './FilterMenu.vue';
 
 const config = useConfigStore();
-const { getModelCapabilities } = useModelList();
 
 // State
 const { 
@@ -36,6 +35,7 @@ const isOpened = ref<boolean>(false);
 // Refs
 const searchBarRef = ref<HTMLInputElement | null>(null);
 const listItemsRef = ref<Array<ComponentPublicInstance<{ listItemRef: HTMLLIElement | null }>>>([]);
+const filterMenu = ref<InstanceType<typeof FilterMenu> | null>(null);
 
 // Lifecycle hooks
 onMounted(async () => {
@@ -181,42 +181,10 @@ function toggleFilterMenu() {
     filterMenuOpen.value = !filterMenuOpen.value;
 }
 
-const orderBy = ref<'default' | 'alphabetically' | 'size'>('default');
-const filterCapabilities = ref<OllamaCapability[]>([]);
-const direction = ref<'asc' | 'des'>('asc');
-
 function userSort(items: ModelInfoListItem[]) {
-    if (filterCapabilities.value.length > 0) {
-        items = items.filter((item) => {
-            const capabilities = getModelCapabilities(item);
-            return filterCapabilities.value.every((cap) => capabilities.includes(cap));
-        });
-    };
-
-    switch (orderBy.value) {
-        case 'default':
-            break;
-        
-        case 'alphabetically':
-            items = items.sort((a, b) => {
-                const item1 = a.modelData.model.split('/')[1] ?? a.modelData.model;
-                const item2 = b.modelData.model.split('/')[1] ?? b.modelData.model;
-
-                return item1.localeCompare(item2, undefined, { sensitivity: 'base' });
-            });
-            break;
-
-        case 'size':
-            items = items.sort((a, b) => a.modelData.size - b.modelData.size);
-            break;
-    }
-
-    if (direction.value === 'des') {
-        items = items.reverse();
-    }
-
-    return items;
+    return filterMenu.value?.userSort(items) || items;
 }
+
 </script>
 
 <template>
@@ -269,62 +237,7 @@ function userSort(items: ModelInfoListItem[]) {
                 </RouterLink>
             </div>
 
-            <div v-if="filterMenuOpen" class="max-h-16 relative flex flex-row gap-2 pb-2 overflow-auto">
-                <div class="flex flex-col justify-end">
-                    <button 
-                        class="bg-surface-light p-2 rounded-md ring-inset ring-2 ring-border-muted h-min"
-                        @click="filterCapabilities = []; orderBy = 'default'; direction = 'asc'">
-                        <BiRefresh class="size-4" />
-                    </button>
-                </div>
-                <label class="flex flex-col">
-                    <span>Filter:</span>
-                    <MultiItemSelect 
-                        v-model="filterCapabilities"
-                        :items="[
-                            {
-                                label: 'Thinking',
-                                value: 'thinking',
-                                icon: BiBrain,
-                            },
-                            {
-                                label: 'Vision',
-                                value: 'vision',
-                                icon: AiOutlineEye,
-                            },
-                            {
-                                label: 'Tools',
-                                value: 'tools',
-                                icon: BiWrench,
-                            }
-                        ]"
-                        button-class="bg-surface-light w-full p-2 rounded-md ring-inset ring-2 ring-border-muted focus:ring-border outline-0 line-clamp-1 "
-                        menu-class="bg-surface w-full min-w-fit p-2 rounded-md ring-inset ring-2 ring-border-muted focus:ring-border outline-0 max-h-48 overflow-y-auto"
-                        item-class="p-1 rounded-md hover:bg-surface-light"
-                        selected-item-class="bg-primary! text-background" />
-                </label>
-                <label class="flex flex-col">
-                    <span>Order:</span>
-                    <select 
-                        v-model="orderBy" 
-                        ref="orderBySelect"
-                        class="bg-surface-light p-2 rounded-md ring-inset ring-2 ring-border-muted focus:ring-border outline-0">
-                        <option value="default">Default ({{ config.api.enabled ? 'Creator Name' : 'Added date' }})</option>
-                        <option value="alphabetically">Alphabetically</option>
-                        <option value="size">Size</option>
-                    </select>
-                </label>
-                <label class="flex flex-col">
-                    <span>Direction:</span>
-                    <select 
-                        v-model="direction" 
-                        ref="directionSelect"
-                        class="bg-surface-light p-2 rounded-md ring-inset ring-2 ring-border-muted focus:ring-border outline-0">
-                        <option value="des">Descending</option>
-                        <option value="asc">Ascending</option>
-                    </select>
-                </label>
-            </div>
+            <FilterMenu ref="filterMenu" :filterMenuOpen />
 
             <ul role="list" :class="{ 'h-62!': filterMenuOpen }" class="h-80 overflow-y-auto [scrollbar-width:thin] *:not-last:mb-2">
                 <li v-if="modelsLoading" class="h-24 flex justify-center items-center">
